Changed the ResultPage constructor to only take named parameters
This is part of Flutter's standard for widgets - all parameters of widgets must be named. Also added constraints to the parameters through the use of the assert method; any parameters passed that do not meet those assertions will throw an error and the entire process will fail. Also added the Triage Grading checkbox.